+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/mapping/MappingSpec.scala
@@ -50,7 +50,7 @@ object MappingSpec extends TypeRegistry[MappingSpec] {
     new JsonSubTypes.Type(name = "case", value = classOf[CaseMappingSpec]),
     new JsonSubTypes.Type(name = "coalesce", value = classOf[CoalesceMappingSpec]),
     new JsonSubTypes.Type(name = "conform", value = classOf[ConformMappingSpec]),
-    new JsonSubTypes.Type(name = "const", value = classOf[ConstMappingSpec]),
+    new JsonSubTypes.Type(name = "const", value = classOf[ValuesMappingSpec]),
     new JsonSubTypes.Type(name = "deduplicate", value = classOf[DeduplicateMappingSpec]),
     new JsonSubTypes.Type(name = "distinct", value = classOf[DistinctMappingSpec]),
     new JsonSubTypes.Type(name = "drop", value = classOf[DropMappingSpec]),
@@ -83,7 +83,8 @@ object MappingSpec extends TypeRegistry[MappingSpec] {
     new JsonSubTypes.Type(name = "union", value = classOf[UnionMappingSpec]),
     new JsonSubTypes.Type(name = "unit", value = classOf[UnitMappingSpec]),
     new JsonSubTypes.Type(name = "unpackJson", value = classOf[UnpackJsonMappingSpec]),
-    new JsonSubTypes.Type(name = "upsert", value = classOf[UpsertMappingSpec])
+    new JsonSubTypes.Type(name = "upsert", value = classOf[UpsertMappingSpec]),
+    new JsonSubTypes.Type(name = "values", value = classOf[ValuesMappingSpec]),
 ))
 abstract class MappingSpec extends NamedSpec[Mapping] {
     @JsonProperty("broadcast") protected var broadcast:String = "false"