+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/target/TargetSpec.scala
@@ -60,8 +60,9 @@ object TargetSpec extends TypeRegistry[TargetSpec] {
     new JsonSubTypes.Type(name = "schema", value = classOf[SchemaTargetSpec]),
     new JsonSubTypes.Type(name = "sftpUpload", value = classOf[SftpUploadTargetSpec]),
     new JsonSubTypes.Type(name = "stream", value = classOf[StreamTargetSpec]),
-    new JsonSubTypes.Type(name = "template", value = classOf[TemplateTargetSpec]))
-)
+    new JsonSubTypes.Type(name = "template", value = classOf[TemplateTargetSpec]),
+    new JsonSubTypes.Type(name = "validate", value = classOf[ValidateTargetSpec])
+))
 abstract class TargetSpec extends NamedSpec[Target] {
     @JsonProperty(value = "before", required=false) private var before:Seq[String] = Seq()
     @JsonProperty(value = "after", required=false) private var after:Seq[String] = Seq()