+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/job/PhaseCommand.scala
@@ -90,6 +90,7 @@ sealed class PhaseCommand(phase:Phase) extends ActionCommand {
     }
 }
 
+class ValidateCommand extends PhaseCommand(Phase.VALIDATE)
 class CreateCommand extends PhaseCommand(Phase.CREATE)
 class BuildCommand extends PhaseCommand(Phase.BUILD)
 class VerifyCommand extends PhaseCommand(Phase.VERIFY)