+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/project/PhaseCommand.scala
@@ -93,12 +93,10 @@ sealed class PhaseCommand(phase:Phase) extends ActionCommand {
     }
 }
 
+
+class ValidateCommand extends PhaseCommand(Phase.VALIDATE)
 class CreateCommand extends PhaseCommand(Phase.CREATE)
-
 class BuildCommand extends PhaseCommand(Phase.BUILD)
-
 class VerifyCommand extends PhaseCommand(Phase.VERIFY)
-
 class TruncateCommand extends PhaseCommand(Phase.TRUNCATE)
-
 class DestroyCommand extends PhaseCommand(Phase.DESTROY)